FarmE-_Blockchain/requirements.txt
uvicorn==0.30.1
python-dotenv==1.0.1
fastapi==0.111.0
supabase==2.6.0
supafunc==0.5.0
ujson==5.10.0
rsa==4.9
